+++ src/Sentry/SentryEvent.cs
@@ -139,12 +139,12 @@ public sealed class SentryEvent : IEventLike, IJsonSerializable
         set => _contexts.ReplaceWith(value);
     }
 
-    private User? _user;
+    private SentryUser? _user;
 
     /// <inheritdoc />
-    public User User
+    public SentryUser User
     {
-        get => _user ??= new User();
+        get => _user ??= new SentryUser();
         set => _user = value;
     }
 
@@ -304,7 +304,7 @@ public sealed class SentryEvent : IEventLike, IJsonSerializable
         var transaction = json.GetPropertyOrNull("transaction")?.GetString();
         var request = json.GetPropertyOrNull("request")?.Pipe(Request.FromJson);
         var contexts = json.GetPropertyOrNull("contexts")?.Pipe(Contexts.FromJson);
-        var user = json.GetPropertyOrNull("user")?.Pipe(User.FromJson);
+        var user = json.GetPropertyOrNull("user")?.Pipe(SentryUser.FromJson);
         var environment = json.GetPropertyOrNull("environment")?.GetString();
         var sdk = json.GetPropertyOrNull("sdk")?.Pipe(SdkVersion.FromJson) ?? new SdkVersion();
         var fingerprint = json.GetPropertyOrNull("fingerprint")?.EnumerateArray().Select(j => j.GetString()).ToArray();