+++ test/Sentry.Extensions.Logging.Tests/SentryLoggingOptionsTests.cs
@@ -5,6 +5,10 @@ namespace Sentry.Extensions.Logging.Tests;
 
 public class SentryLoggingOptionsTests : BindableTests<SentryLoggingOptions>
 {
+    public SentryLoggingOptionsTests() : base(nameof(SentryLoggingOptions.ExperimentalMetrics))
+    {
+    }
+
     [Fact]
     public void BindableProperties_MatchOptionsProperties()
     {